算法稽核 10 單位修到好結案:Scala 凍結碼 + SQL view + q01(全專案 sbt 編譯綠 + Python 測試零回歸)
audit-closure workflow(10 agent 平行,worktree 各自 sbt compile/sqlglot/py_compile 自驗)產出
96 個精確 edits,主 loop 統一套用 + 全專案 sbt compile(exit 0、僅既有 deprecation)+ Python
62 測試綠整合。逐單位:

Scala 資料層(退役前修到好結案,live 已改吃 Python;仍要編譯通過):
- slick-schema:DailyTradingDetails 22 個法人買賣超股數欄 Int→Long(單日單檔可達數十億股,
  Int 上限 2.1e9 超界會 NumberFormatException 靜默寫 0/回捲;PG 實測已用到 43%)+ parser
  transferValues 改 toLong;NetChangeOfPrice 投影漏「漲停家數」補回;ExRightDividend cash_dividend
  加語意註解(實存權值+息值總調整額、跨 2024-07 翻純現金)。DB 端另需 ALTER COLUMN bigint 維持
  FRM parity(現值未溢位故非 breaking,PG 退役後 moot)。
- crawler-scala:休市 sentinel 判定加正向交易日證據(舊版凡 <50 bytes 就無條件當休市,害 27 個
  真交易日 twse 資料被永久標休市丟棄)。
- quality-screening:QualityFilter 加 DISTINCT ON 收斂到 PIT 最新一季(舊版「歷史任一季曾達標
  就永久放行」= 40~49% 通過名單其實已惡化)。

Scala 策略/指標(凍結參考碼):
- metrics-scala:Sortino 下行差改全期 N + MAR 錨、Sharpe 分子改算術年化超額、turnover 分母改逐期
  NAV 均值、TDPY 242 註出處。
- rankmetrics-scala:forwardReturns 未來報酬改還原價(舊用原始收盤 → 除息旺季系統性壓低價值/
  殖利率因子 IC);FactorResearch 同缺陷複本一併修。
- signals-scala:distFrom52wHigh 的 52 週窗改日曆年(舊 INTERVAL 252 days 把交易日當日曆天=只看
  8 個月);Greenblatt EBIT/EV 負債/投入資本口徑修正對齊 docstring 宣稱定義。
- backtester-scala:載入 capital_reduction 表、減資日等比調整股數/還原參考價(舊版 386 筆減資 91%
  漏接,把會計性跳價當真漲幅 → 系統性膨脹報酬,實測 8103 憑空 +11.6%)。

SQL view(PG-only、Python raw_quarterly 為唯一真源,隨 PG 退役封存,但先修對):
- sql-finratios:TTM 總資產週轉率分母 lag5→lag4(年初資產打字錯)、現金比率分母改流動負債、
  ROIC 分子改營業利益、負權益/負資本運用 rank guard。
- sql-valuation:五線譜通道寬改用迴歸殘差標準差(非收盤價原始 std)、DCF 成長率改幾何 CAGR、
  eps_growth_rate_10y 視窗排序修正。

研究(q01 IC 廣篩,非生產):
- q01-bankruptcy:Altman Z'' X2 用真保留盈餘(非「權益−股本」代理,舊版抹掉 114 家累虧訊號)、
  Ohlson INTWO/CHIN 改連兩「年」(shift4,非連兩季)、X3 EBIT 用稅前淨利代理。raw_quarterly.py
  加 pretax/retained_earnings 兩科目(純加法、不動既有 F-Score 欄;S live 不受影響)。

驗證:全專案 sbt compile exit 0(90 sources);strat_lab+apex 62 測試綠(raw_quarterly 加欄未破
F-Score/S);SQL 經 sqlglot(postgres)+DuckDB 解析 + turnover 5→4 spot-check 確認。

+++ src/main/scala/reader/TradingReader.scala
@@ -183,9 +183,9 @@ class TradingReader extends Reader {
             rows.map {
               values =>
                 val splitValues = values.splitAt(2)
-                val transferValues = splitValues._2.init.map(value => Try(value.toInt).getOrElse(0))
+                val transferValues = splitValues._2.init.map(value => Try(value.toLong).getOrElse(0L))
                 val companyCode = values.head
-                val noneInt: Option[Int] = None
+                val noneInt: Option[Long] = None
                 values.size match {
                   case 13 =>
                     marketFile.market :: date :: companyCode :: values(1) :: noneInt :: noneInt :: noneInt :: noneInt :: noneInt :: noneInt ::
@@ -211,9 +211,9 @@ class TradingReader extends Reader {
             rows.map {
               values =>
                 val splitValues = values.splitAt(2)
-                val transferValues = splitValues._2.map(value => Try(value.toInt).getOrElse(0))
+                val transferValues = splitValues._2.map(value => Try(value.toLong).getOrElse(0L))
                 val companyCode = values.head
-                val noneInt: Option[Int] = None
+                val noneInt: Option[Long] = None
                 values.size match {
                   case 12 =>
                     marketFile.market :: date :: companyCode :: values(1) ::
